refactor: move SQL driver registration into pkg/driver
Move driver_{mysql,pgsql,sqlite3}.go and driver_mysql_test.go out of the
root package into a dedicated pkg/driver package so the driver wiring is
organized in one place.

The MySQL DSN normalizer previously assigned an unexported
normalizeMySQLDSN hook in db.go from driver_mysql.go's init. Moving the
driver to its own package would have required pkg/driver to import the
root package while the root package imports pkg/driver — an import cycle.
To avoid it, ownership of the hook now lives in pkg/driver as the
exported NormalizeMySQLDSN var; db.go reads it, giving a single
dependency direction (rockhopper -> pkg/driver). db.go's import of
pkg/driver also auto-registers the bundled drivers, preserving the
existing behavior the test suite relies on.

An unconstrained pkg/driver/driver.go declares NormalizeMySQLDSN and
keeps the package compiling when every driver is excluded via the
no_mysql/no_postgres/no_sqlite3 build tags.
