+++ pkg/driver/pgsql.go
@@ -1,7 +1,7 @@
 //go:build !no_postgres
 // +build !no_postgres
 
-package rockhopper
+package driver
 
 import (
 	_ "github.com/lib/pq"
