+++ pkg/driver/sqlite3.go
@@ -1,7 +1,7 @@
 //go:build !no_sqlite3
 // +build !no_sqlite3
 
-package rockhopper
+package driver
 
 import (
 	_ "github.com/mattn/go-sqlite3"
